Routing is a thing now
So I’m stupid, and in moving stuff around to fit the style guide
(moving things into their own files), I broke Gulp a little bit. It was
gulping things in alphabetical order, so aboutController was trying to
find the module before it was made. Hence the lazy fix of renaming the
main file to aAngularApp :D Feel free to fix my shit if you want, or I
can fix it in the future.