Images!
These are from the alpha sig image library. a little colorization and
hue manipulation and viola! The red may not be the exact shade we want,
but we can always adjust it if need be. Also, the phoenix images are
crazy tall so some readjustment on those might be necessary as well. Let
me know what you need and I'll get it done!